# api-client_generated_api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Adversarial Candidate Ranking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  fullName: string;
  company: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  company: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface JobInput {
  title: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  experienceRequired?: number | null;
  skillsRequired?: string[];
  rawText: string;
}

export interface JobUpdate {
  title?: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  experienceRequired?: number | null;
  skillsRequired?: string[];
  rawText?: string;
}

export interface JobDescription {
  id: number;
  title: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  experienceRequired?: number | null;
  skillsRequired?: string[];
  rawText: string;
  createdAt: string;
}

export interface CandidateInput {
  name: string;
  /** @nullable */
  email?: string | null;
  skills?: string[];
  /** @nullable */
  experienceYears?: number | null;
  /** @nullable */
  education?: string | null;
  rawText: string;
}

export interface Candidate {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  skills?: string[];
  /** @nullable */
  experienceYears?: number | null;
  /** @nullable */
  education?: string | null;
  rawText: string;
  createdAt: string;
}

export interface RankingStartInput {
  jdId: number;
  candidateIds: number[];
}

export type RankingJobStatus = typeof RankingJobStatus[keyof typeof RankingJobStatus];


export const RankingJobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface RankingJob {
  id: number;
  jdId?: number;
  status: RankingJobStatus;
  progress: number;
  totalCandidates: number;
  /** @nullable */
  jobTitle?: string | null;
  createdAt: string;
}

export type RankedCandidateManipulationRisk = typeof RankedCandidateManipulationRisk[keyof typeof RankedCandidateManipulationRisk];


export const RankedCandidateManipulationRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type RankedCandidateRecommendation = typeof RankedCandidateRecommendation[keyof typeof RankedCandidateRecommendation];


export const RankedCandidateRecommendation = {
  strong_hire: 'strong_hire',
  hire: 'hire',
  consider: 'consider',
  reject: 'reject',
} as const;

export type RedFlagSeverity = typeof RedFlagSeverity[keyof typeof RedFlagSeverity];


export const RedFlagSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface RedFlag {
  type: string;
  severity: RedFlagSeverity;
  details: string;
}

export interface SkillsMatch {
  matched: string[];
  missing: string[];
  extra: string[];
}

export interface AdversarialTest {
  perturbationType: string;
  originalRank: number;
  perturbedRank: number;
  positionChange: number;
}

export interface RankedCandidate {
  candidateId: number;
  name: string;
  finalScore: number;
  fitScore: number;
  robustnessScore: number;
  stabilityScore: number;
  ranking: number;
  isRobust: boolean;
  manipulationRisk?: RankedCandidateManipulationRisk;
  redFlags: RedFlag[];
  skillsMatch?: SkillsMatch;
  explanation: string;
  recommendation: RankedCandidateRecommendation;
  strengths?: string[];
  gaps?: string[];
  adversarialTests?: AdversarialTest[];
}

export interface RankingSummary {
  totalCandidates: number;
  flaggedForManipulation: number;
  highlyRobust: number;
  averageRobustness: number;
  timeTaken: string;
}

export interface RankingResults {
  jobId: number;
  rankedCandidates: RankedCandidate[];
  summary: RankingSummary;
}

export type AdversaryConfigInputIntensity = typeof AdversaryConfigInputIntensity[keyof typeof AdversaryConfigInputIntensity];


export const AdversaryConfigInputIntensity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AdversaryConfigInput {
  numVariants?: number;
  intensity?: AdversaryConfigInputIntensity;
  enabledPerturbations?: string[];
}

export type AdversaryConfigIntensity = typeof AdversaryConfigIntensity[keyof typeof AdversaryConfigIntensity];


export const AdversaryConfigIntensity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AdversaryConfig {
  id?: number;
  numVariants: number;
  intensity: AdversaryConfigIntensity;
  enabledPerturbations: string[];
}

export interface PerturbationType {
  key: string;
  label: string;
  description: string;
}

export interface DashboardStats {
  totalCandidates: number;
  totalJobs: number;
  totalRankings: number;
  flaggedProfiles: number;
  highlyRobust: number;
  averageScore: number;
  recentRankings?: RankingJob[];
}

export interface TrendDataPoint {
  date: string;
  totalRankings: number;
  flaggedCount: number;
  robustCount: number;
  averageScore: number;
}

export interface ManipulationStats {
  keywordStuffing: number;
  experienceInflation: number;
  promptInjection: number;
  irrelevantSkills: number;
  total: number;
}

export type ListCandidatesParams = {
jobId?: number;
};