# api-spec_openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Adversarial Candidate Ranking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job descriptions
      responses:
        "200":
          description: List of job descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobDescription"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescription"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job description by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescription"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update job description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDescription"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # CANDIDATES
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # RANKING
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
    RegisterInput:
      type: object
      required: [email, password, fullName, company]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        company:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, fullName, company, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        company:
          type: string
        role:
          type: string
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # JOBS
    JobInput:
      type: object
      required: [title, rawText]
      properties:
        title:
          type: string
        department:
          type: ["string", "null"]
        experienceRequired:
          type: ["integer", "null"]
        skillsRequired:
          type: array
          items:
            type: string
        rawText:
          type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        department:
          type: ["string", "null"]
        experienceRequired:
          type: ["integer", "null"]
        skillsRequired:
          type: array
          items:
            type: string
        rawText:
          type: string
    JobDescription:
      type: object
      required: [id, title, rawText, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: ["string", "null"]
        experienceRequired:
          type: ["integer", "null"]
        skillsRequired:
          type: array
          items:
            type: string
        rawText:
          type: string
        createdAt:
          type: string

    # CANDIDATES
    CandidateInput:
      type: object
      required: [name, rawText]
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: ["integer", "null"]
        education:
          type: ["string", "null"]
        rawText:
          type: string
    Candidate:
      type: object
      required: [id, name, rawText, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: ["integer", "null"]
        education:
          type: ["string", "null"]
        rawText:
          type: string
        createdAt:
          type: string

    # RANKING
...